requirements.txt
# Discord Bot TDD Implementation - Requirements
# Python 3.8+ required

# Core Discord Bot framework
discord.py>=2.3.0

# OpenAI API for GPT-4o-mini and Whisper
openai>=0.27.8

# Redis for rate limiting
redis>=4.5.0

# Environment variables
python-dotenv>=1.0.0

# For testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
fakeredis>=2.21.0
moto[s3]>=5.0.0
pytest-httpx>=0.27.0

# Audio/video processing (requires ffmpeg system installation)
# Note: ffmpeg must be installed separately on the system

# File processing
pdfminer.six>=20221105  # PDF text extraction

# Utilities
aiofiles>=23.0.0
typing-extensions>=4.5.0